refactor: Jwt 토큰 추출하는 클래스 이름 변경
